+++ apps/web/src/components/Notification/List.tsx
@@ -20,7 +20,7 @@ import { type FC } from 'react';
 import { Virtuoso } from 'react-virtuoso';
 import { NotificationTabType } from 'src/enums';
 import { usePreferencesStore } from 'src/store/non-persisted/usePreferencesStore';
-import { useNotificationPersistStore } from 'src/store/useNotificationPersistStore';
+import { useNotificationPersistStore } from 'src/store/persisted/useNotificationPersistStore';
 import { useUpdateEffect } from 'usehooks-ts';
 
 import NotificationShimmer from './Shimmer';
