+++ apps/web/src/store/persisted/useEnabledAlgorithmsPersistStore.ts
